adds start of debug menu
Use CTRL+M to open to open the debug menu. You can make the game not
start on the debug menu by changing "Debug menu on start" to false in
The Betweenlands configuration file.
- slight render code cleaning up

+++ java/thebetweenlands/client/render/block/BlockMudFlowerPotRenderer.java
@@ -1,8 +1,6 @@
 package thebetweenlands.client.render.block;
 
-import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
 import net.minecraft.block.Block;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.RenderBlocks;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.item.Item;
@@ -10,13 +8,11 @@ import net.minecraft.item.ItemBlock;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.tileentity.TileEntityFlowerPot;
 import net.minecraft.world.IBlockAccess;
-import net.minecraft.world.World;
-import org.lwjgl.opengl.GL11;
 import thebetweenlands.blocks.BLBlockRegistry;
-import thebetweenlands.blocks.BlockBLFlowerPot;
 import thebetweenlands.client.model.block.ModelMudFlowerPot;
 import thebetweenlands.proxy.ClientProxy;
 import thebetweenlands.utils.ModelConverter;
+import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
 
 /**
  * Created by Bart on 21-6-2015.
@@ -70,22 +66,22 @@ public class BlockMudFlowerPotRenderer implements ISimpleBlockRenderingHandler {
                 l = flower.colorMultiplier(world, x, y, z);
 
                 if (l != 16777215) {
-                    float f = (float) (l >> 16 & 255) / 255.0F;
-                    float f1 = (float) (l >> 8 & 255) / 255.0F;
-                    float f2 = (float) (l & 255) / 255.0F;
+                    float f = (l >> 16 & 255) / 255.0F;
+                    float f1 = (l >> 8 & 255) / 255.0F;
+                    float f2 = (l & 255) / 255.0F;
                     tessellator.setColorOpaque_F(f, f1, f2);
                 }
 
                 if (j1 == 1) {
-                    renderer.drawCrossedSquares(renderer.getBlockIconFromSideAndMetadata(flower, 0, i1), (double) x, (double) y, (double) z, 0.75F);
+                    renderer.drawCrossedSquares(renderer.getBlockIconFromSideAndMetadata(flower, 0, i1), x, y, z, 0.75F);
                 } else if (j1 == 13) {
                     renderer.renderAllFaces = true;
                     float f9 = 0.125F;
-                    renderer.setRenderBounds((double) (0.5F - f9), 0.0D, (double) (0.5F - f9), (double) (0.5F + f9), 0.25D, (double) (0.5F + f9));
+                    renderer.setRenderBounds(0.5F - f9, 0.0D, 0.5F - f9, 0.5F + f9, 0.25D, 0.5F + f9);
                     renderer.renderStandardBlock(flower, x, y, z);
-                    renderer.setRenderBounds((double) (0.5F - f9), 0.25D, (double) (0.5F - f9), (double) (0.5F + f9), 0.5D, (double) (0.5F + f9));
+                    renderer.setRenderBounds(0.5F - f9, 0.25D, 0.5F - f9, 0.5F + f9, 0.5D, 0.5F + f9);
                     renderer.renderStandardBlock(flower, x, y, z);
-                    renderer.setRenderBounds((double) (0.5F - f9), 0.5D, (double) (0.5F - f9), (double) (0.5F + f9), 0.75D, (double) (0.5F + f9));
+                    renderer.setRenderBounds(0.5F - f9, 0.5D, 0.5F - f9, 0.5F + f9, 0.75D, 0.5F + f9);
                     renderer.renderStandardBlock(flower, x, y, z);
                     renderer.renderAllFaces = false;
                     renderer.setRenderBounds(0.0D, 0.0D, 0.0D, 1.0D, 1.0D, 1.0D);
